refactor: GIFContent 프로토콜로 변경
