+++ GIFApp/Domain/Content/GIFItem.swift
@@ -1,0 +1,163 @@
+//
+//  GIFContent.swift
+//  GIFApp
+//
+//  Created by 김동현 on 2023/05/24.
+//
+
+import Foundation
+import UIKit
+
+struct GIFItem: GIFContent {    
+    var id: String?
+    var path: String?
+    var owner: String?
+    var bannerPath: String?
+    
+    var convertedGif: UIImage? {
+        return gifImageWithURL(self.path ?? "")
+    }
+}
+
+extension GIFItem {
+    init(_ item: GifResponse.ImageData?) {
+        self.init(id: item?.images?.original?.hash,
+                  path: item?.images?.original?.url,
+                  owner: item?.user?.userName,
+                  bannerPath: item?.user?.bannerURL)
+    }
+}
+
+extension GIFItem {
+    private func gifImageWithURL(_ gifUrl:String) -> UIImage? {
+        guard let bundleURL = NSURL(string: gifUrl)
+        else {
+            print("image named \"\(gifUrl)\" doesn't exist")
+            return nil
+        }
+        guard let imageData = NSData(contentsOf: bundleURL as URL) else {
+            print("image named \"\(gifUrl)\" into NSData")
+            return nil
+        }
+        
+        return gifImageWithData(data: imageData)
+    }
+    
+    private func gifImageWithData(data: NSData) -> UIImage? {
+        guard let source = CGImageSourceCreateWithData(data, nil) else {
+            print("image doesn't exist")
+            return nil
+        }
+        
+        return self.animatedImageWithSource(source: source)
+    }
+    private func animatedImageWithSource(source: CGImageSource) -> UIImage? {
+        let count = CGImageSourceGetCount(source)
+        var images = [CGImage]()
+        var delays = [Int]()
+        
+        for i in 0..<count {
+            if let image = CGImageSourceCreateImageAtIndex(source, i, nil) {
+                images.append(image)
+            }
+            
+            let delaySeconds = self.delayForImageAtIndex(index: Int(i), source: source)
+            delays.append(Int(delaySeconds * 1000.0)) // Seconds to ms
+        }
+        
+        let duration: Int = {
+            var sum = 0
+            
+            for val: Int in delays {
+                sum += val
+            }
+            
+            return sum
+        }()
+        
+        let gcd = gcdForArray(array: delays)
+        var frames = [UIImage]()
+        
+        var frame: UIImage
+        var frameCount: Int
+        for i in 0..<count {
+            frame = UIImage(cgImage: images[Int(i)])
+            frameCount = Int(delays[Int(i)] / gcd)
+            
+            for _ in 0..<frameCount {
+                frames.append(frame)
+            }
+        }
+        
+        let animation = UIImage.animatedImage(with: frames, duration: Double(duration) / 1000.0)
+        
+        return animation
+    }
+    
+    private func delayForImageAtIndex(index: Int, source: CGImageSource!) -> Double {
+        var delay = 0.1
+        
+        let cfProperties = CGImageSourceCopyPropertiesAtIndex(source, index, nil)
+        let gifProperties: CFDictionary = unsafeBitCast(CFDictionaryGetValue(cfProperties, Unmanaged.passUnretained(kCGImagePropertyGIFDictionary).toOpaque()), to: CFDictionary.self)
+        
+        var delayObject: AnyObject = unsafeBitCast(CFDictionaryGetValue(gifProperties, Unmanaged.passUnretained(kCGImagePropertyGIFUnclampedDelayTime).toOpaque()), to: AnyObject.self)
+        
+        if delayObject.doubleValue == 0 {
+            delayObject = unsafeBitCast(CFDictionaryGetValue(gifProperties, Unmanaged.passUnretained(kCGImagePropertyGIFDelayTime).toOpaque()), to: AnyObject.self)
+        }
+        
+        delay = delayObject as! Double
+        
+        if delay < 0.1 {
+            delay = 0.1
+        }
+        
+        return delay
+    }
+    
+    private func gcdForArray(array: Array<Int>) -> Int {
+        if array.isEmpty {
+            return 1
+        }
+        
+        var gcd = array[0]
+        
+        for val in array {
+            gcd = self.gcdForPair(a: val, gcd)
+        }
+        
+        return gcd
+    }
+    
+    private func gcdForPair(a: Int?, _ b: Int?) -> Int {
+        var a = a
+        var b = b
+        if b == nil || a == nil {
+            if b != nil {
+                return b!
+            } else if a != nil {
+                return a!
+            } else {
+                return 0
+            }
+        }
+        
+        if a! < b! {
+            let c = a!
+            a = b!
+            b = c
+        }
+        
+        var rest: Int
+        while true {
+            rest = a! % b!
+            
+            if rest == 0 {
+                return b!
+            } else {
+                a = b!
+                b = rest
+            }
+        }
+    }
+}
